% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config_ini.R
\name{check_config_ini}
\alias{check_config_ini}
\title{Check Monolix/Simulx config file}
\usage{
check_config_ini(
  cmd = getOption("mlxtrx.monolix_cmd", "mono24"),
  offer_fix = interactive()
)
}
\arguments{
\item{cmd}{Monolix or Simulx command being executed. Defaults to \code{"mono24"}}

\item{offer_fix}{Whether to offer interactive fixes. Defaults to \code{interactive()}}
}
\value{
Invisibly returns TRUE if all settings are present
}
\description{
Checks if the relevant config.ini file contains required export settings
for charts data.
In interactive sessions, offers to fix missing settings.
}
\examples{
\dontrun{
check_config_ini()
}
}
