% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono.R
\name{execute_job}
\alias{execute_job}
\title{Execute a single Monolix job}
\usage{
execute_job(
  path,
  output_dir = NULL,
  thread = NULL,
  tool = NULL,
  mode = NULL,
  config = NULL,
  cmd = "mono24"
)
}
\arguments{
\item{path}{Character scalar. Path to a single existing Monolix project file.}

\item{output_dir}{Character scalar or \code{NULL}. Directory for output files.}

\item{thread}{Integer scalar or \code{NULL}. Number of threads used by Monolix.}

\item{tool}{Character scalar or \code{NULL}. Tool to launch assessment.}

\item{mode}{Character scalar or \code{NULL}. Console mode.}

\item{config}{Character scalar or \code{NULL}. Configuration file path.}

\item{cmd}{Character scalar. The Monolix command to execute.}
}
\value{
Integer scalar with job ID, or \code{NA} if job submission failed.
}
\description{
Submits a single Monolix job to the grid and extracts the job ID.
}
\keyword{internal}
