% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{FileHandler}
\alias{FileHandler}
\title{Base file handler class}
\description{
Base file handler class

Base file handler class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileHandler-parse_file}{\code{FileHandler$parse_file()}}
\item \href{#method-FileHandler-get_input_files}{\code{FileHandler$get_input_files()}}
\item \href{#method-FileHandler-get_output_dir}{\code{FileHandler$get_output_dir()}}
\item \href{#method-FileHandler-can_monitor}{\code{FileHandler$can_monitor()}}
\item \href{#method-FileHandler-clone}{\code{FileHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHandler-parse_file"></a>}}
\if{latex}{\out{\hypertarget{method-FileHandler-parse_file}{}}}
\subsection{Method \code{parse_file()}}{
Parse project file and extract metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHandler$parse_file(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHandler-get_input_files"></a>}}
\if{latex}{\out{\hypertarget{method-FileHandler-get_input_files}{}}}
\subsection{Method \code{get_input_files()}}{
Get input files from parsed content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHandler$get_input_files(parsed_content, project_dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parsed_content}}{Result from parse_file()}

\item{\code{project_dir}}{Directory containing the project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHandler-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-FileHandler-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
Get expected output directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHandler$get_output_dir(path, output_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}

\item{\code{output_dir}}{User-specified output directory (may be NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHandler-can_monitor"></a>}}
\if{latex}{\out{\hypertarget{method-FileHandler-can_monitor}{}}}
\subsection{Method \code{can_monitor()}}{
Check if job monitoring is supported
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHandler$can_monitor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
