% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config_ini.R
\name{fix_ini_settings}
\alias{fix_ini_settings}
\title{Fix INI file settings by adding missing and replacing incorrect ones}
\usage{
fix_ini_settings(ini_lines, missing_settings, incorrect_settings)
}
\arguments{
\item{ini_lines}{Original INI file lines}

\item{missing_settings}{Vector of missing settings to add}

\item{incorrect_settings}{Vector of settings to replace}
}
\value{
Modified INI file lines
}
\description{
Fix INI file settings by adding missing and replacing incorrect ones
}
\keyword{internal}
