% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{GenericHandler}
\alias{GenericHandler}
\title{Generic file handler for unknown file types}
\description{
Generic file handler for unknown file types

Generic file handler for unknown file types
}
\keyword{internal}
\section{Super class}{
\code{\link[mlxtrx:FileHandler]{mlxtrx::FileHandler}} -> \code{GenericHandler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenericHandler-parse_file}{\code{GenericHandler$parse_file()}}
\item \href{#method-GenericHandler-get_input_files}{\code{GenericHandler$get_input_files()}}
\item \href{#method-GenericHandler-get_output_dir}{\code{GenericHandler$get_output_dir()}}
\item \href{#method-GenericHandler-can_monitor}{\code{GenericHandler$can_monitor()}}
\item \href{#method-GenericHandler-clone}{\code{GenericHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericHandler-parse_file"></a>}}
\if{latex}{\out{\hypertarget{method-GenericHandler-parse_file}{}}}
\subsection{Method \code{parse_file()}}{
Parse project file and extract metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHandler$parse_file(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericHandler-get_input_files"></a>}}
\if{latex}{\out{\hypertarget{method-GenericHandler-get_input_files}{}}}
\subsection{Method \code{get_input_files()}}{
Get input files from parsed content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHandler$get_input_files(parsed_content, project_dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parsed_content}}{Result from parse_file()}

\item{\code{project_dir}}{Directory containing the project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericHandler-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-GenericHandler-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
Get expected output directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHandler$get_output_dir(path, output_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}

\item{\code{output_dir}}{User-specified output directory (may be NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericHandler-can_monitor"></a>}}
\if{latex}{\out{\hypertarget{method-GenericHandler-can_monitor}{}}}
\subsection{Method \code{can_monitor()}}{
Check if job monitoring is supported
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHandler$can_monitor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenericHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
