% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{get_file_handler}
\alias{get_file_handler}
\title{Get appropriate file handler for a given file extension}
\usage{
get_file_handler(ext)
}
\arguments{
\item{ext}{File extension (without dot)}
}
\description{
Get appropriate file handler for a given file extension
}
\keyword{internal}
