% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{MonolixHandler}
\alias{MonolixHandler}
\title{Monolix file handler}
\description{
Monolix file handler

Monolix file handler
}
\keyword{internal}
\section{Super class}{
\code{\link[mlxtrx:FileHandler]{mlxtrx::FileHandler}} -> \code{MonolixHandler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MonolixHandler-parse_file}{\code{MonolixHandler$parse_file()}}
\item \href{#method-MonolixHandler-get_input_files}{\code{MonolixHandler$get_input_files()}}
\item \href{#method-MonolixHandler-get_output_dir}{\code{MonolixHandler$get_output_dir()}}
\item \href{#method-MonolixHandler-parse_summary}{\code{MonolixHandler$parse_summary()}}
\item \href{#method-MonolixHandler-clone}{\code{MonolixHandler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlxtrx" data-topic="FileHandler" data-id="can_monitor"><a href='../../mlxtrx/html/FileHandler.html#method-FileHandler-can_monitor'><code>mlxtrx::FileHandler$can_monitor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlxtrx" data-topic="FileHandler" data-id="get_summary_file_path"><a href='../../mlxtrx/html/FileHandler.html#method-FileHandler-get_summary_file_path'><code>mlxtrx::FileHandler$get_summary_file_path()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonolixHandler-parse_file"></a>}}
\if{latex}{\out{\hypertarget{method-MonolixHandler-parse_file}{}}}
\subsection{Method \code{parse_file()}}{
Parse project file and extract metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonolixHandler$parse_file(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonolixHandler-get_input_files"></a>}}
\if{latex}{\out{\hypertarget{method-MonolixHandler-get_input_files}{}}}
\subsection{Method \code{get_input_files()}}{
Get input files from parsed content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonolixHandler$get_input_files(parsed_content, project_dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parsed_content}}{Result from parse_file()}

\item{\code{project_dir}}{Directory containing the project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonolixHandler-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-MonolixHandler-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
Get expected output directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonolixHandler$get_output_dir(path, output_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}

\item{\code{output_dir}}{User-specified output directory (may be NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonolixHandler-parse_summary"></a>}}
\if{latex}{\out{\hypertarget{method-MonolixHandler-parse_summary}{}}}
\subsection{Method \code{parse_summary()}}{
Parse Monolix summary file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonolixHandler$parse_summary(summary_file_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_file_path}}{Path to summary.txt file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonolixHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MonolixHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonolixHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
