% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config_ini.R
\name{parse_ini_file}
\alias{parse_ini_file}
\title{Parse ini file into structured list}
\usage{
parse_ini_file(ini_path)
}
\arguments{
\item{ini_path}{Path to ini file}
}
\value{
Named list with sections as top-level elements
}
\description{
Parse ini file into structured list
}
\keyword{internal}
