% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print_report_summary}
\alias{print_report_summary}
\title{Print project report summary}
\usage{
print_report_summary(
  directory = ".",
  run_ids = NULL,
  ext = c("mlxtran", "smlx"),
  db_conn = default_db_conn()
)
}
\arguments{
\item{directory}{Character scalar. Directory path to check for project files.}

\item{run_ids}{Integer vector. Specific run IDs to check. If NULL (default),
checks all runs.}

\item{ext}{Character vector. File extensions to consider as
project files. Defaults to c("mlxtran", "smlx").}

\item{db_conn}{A database connection object inheriting from \code{DBIObject}.
Defaults to a connection created by \code{default_db_conn()}.}
}
\value{
NULL (invisible). Called for side effects (printing).
}
\description{
Prints a formatted summary of the project report to the console.
}
