% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config_ini.R
\name{show_ini_diff}
\alias{show_ini_diff}
\title{Show differences between original and modified INI files}
\usage{
show_ini_diff(
  original_lines,
  modified_lines,
  missing_settings,
  incorrect_settings
)
}
\arguments{
\item{original_lines}{Original INI lines}

\item{modified_lines}{Modified INI lines}

\item{missing_settings}{Missing settings that were added}

\item{incorrect_settings}{Incorrect settings that were replaced}
}
\description{
Show differences between original and modified INI files
}
\keyword{internal}
