% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{SimulxHandler}
\alias{SimulxHandler}
\title{Simulx file handler}
\description{
Simulx file handler

Simulx file handler
}
\keyword{internal}
\section{Super class}{
\code{\link[mlxtrx:FileHandler]{mlxtrx::FileHandler}} -> \code{SimulxHandler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulxHandler-parse_file}{\code{SimulxHandler$parse_file()}}
\item \href{#method-SimulxHandler-get_input_files}{\code{SimulxHandler$get_input_files()}}
\item \href{#method-SimulxHandler-get_output_dir}{\code{SimulxHandler$get_output_dir()}}
\item \href{#method-SimulxHandler-clone}{\code{SimulxHandler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlxtrx" data-topic="FileHandler" data-id="can_monitor"><a href='../../mlxtrx/html/FileHandler.html#method-FileHandler-can_monitor'><code>mlxtrx::FileHandler$can_monitor()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulxHandler-parse_file"></a>}}
\if{latex}{\out{\hypertarget{method-SimulxHandler-parse_file}{}}}
\subsection{Method \code{parse_file()}}{
Parse project file and extract metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulxHandler$parse_file(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulxHandler-get_input_files"></a>}}
\if{latex}{\out{\hypertarget{method-SimulxHandler-get_input_files}{}}}
\subsection{Method \code{get_input_files()}}{
Get input files from parsed content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulxHandler$get_input_files(parsed_content, project_dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parsed_content}}{Result from parse_file()}

\item{\code{project_dir}}{Directory containing the project file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulxHandler-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-SimulxHandler-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
Get expected output directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulxHandler$get_output_dir(path, output_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to project file}

\item{\code{output_dir}}{User-specified output directory (may be NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulxHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimulxHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulxHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
