% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-utils.R
\name{resolve_output_dir}
\alias{resolve_output_dir}
\title{Resolve output directory for a Monolix project}
\usage{
resolve_output_dir(path, output_dir = NULL)
}
\arguments{
\item{path}{Character scalar. Path to Monolix project file.}

\item{output_dir}{Character scalar or NULL. Override output directory.}
}
\value{
Character scalar. Resolved output directory path.
}
\description{
Resolve output directory for a Monolix project
}
\keyword{internal}
